Refactor: Extract browser driver startup from watch_media
