+++ src/player/watch_media.rs
@@ -1,4 +1,3 @@
-use std::{process::Command, process::Stdio, thread::sleep, time::Duration};
 use thirtyfour::prelude::*;
 
 use crate::{
@@ -6,48 +5,26 @@ use crate::{
         choose_episode::choose_episode, choose_episode_with_images::choose_episode_with_images,
         choose_lang::choose_lang, choose_season::choose_season,
     },
+    driver::start_driver::{get_driver, start_browser_driver},
     fs::posters::get_posters_path,
     media::Media,
-    player::mpv::open_mpv,
-    player::vlc::open_vlc,
-    TRANSLATION, USE_GECKODRIVER, USE_MPV,
+    player::{mpv::open_mpv, vlc::open_vlc},
+    TRANSLATION, USE_MPV,
 };
 
 #[tokio::main]
 pub async fn watch_media(media: Media, img_mode: bool) -> WebDriverResult<()> {
     let language = TRANSLATION.get().unwrap();
     let use_mpv = USE_MPV.get().unwrap();
-    let use_geckodriver = USE_GECKODRIVER.get().unwrap();
-
-    let url = format!("https://vizer.in/{}", &media.url);
-    let mut driver_command = String::new();
-    if *use_geckodriver {
-        driver_command.push_str("geckodriver");
-    } else {
-        driver_command.push_str("chromedriver");
-    };
-
-    let mut browser_driver = Command::new(driver_command)
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-        .unwrap();
-    // we need to wait command to start :(
-    sleep(Duration::from_millis(100));
 
     print!("\x1B[2J\x1B[1;1H");
     println!("{}", language.preparing_misc_text);
 
-    let driver: WebDriver = if *use_geckodriver {
-        let mut caps = DesiredCapabilities::firefox();
-        caps.set_headless().unwrap();
-        WebDriver::new("http://localhost:4444", caps).await?
-    } else {
-        let mut caps = DesiredCapabilities::chrome();
-        caps.set_headless().unwrap();
-        WebDriver::new("http://localhost:9515", caps).await?
-    };
+    let mut browser_driver = start_browser_driver();
 
+    let driver = get_driver().await;
+
+    let url = format!("https://vizer.in/{}", &media.url);
     driver.goto(url).await?;
 
     if media.url.contains("serie/") {
